add webui_log toggle switch

+++ webUI/app/view/status_view.py
@@ -78,10 +78,11 @@ def getPrjInfo():
 def delete(prj_id):
     # delete records in database
     prj = db.session.query(models.Project).get(prj_id)
-    gene_amount = db.session.query(models.GeneAmount).get(prj.gene_amount)
-    db.session.delete(gene_amount)
-    db.session.delete(prj)
-    db.session.commit()
+    if prj and prj.gene_amount:
+        gene_amount = db.session.query(models.GeneAmount).get(prj.gene_amount)
+        db.session.delete(gene_amount)
+        db.session.delete(prj)
+        db.session.commit()
     # delete project folder
     proj_folder=PRJ_HOME + '/' + prj.id
     if isdir(proj_folder):
